updated page structure
changed some formatting, + added seperate pages for bmi and exercises & spinner
