+++ client/src/components/Navagation.js
@@ -1,5 +1,4 @@
 import React from "react";
-
 function Navigation({ handleLoginClick }) {
 
     const handleClick = () => {
